move makecred and wrap to utils
makecred.py contains more then makecred, so should be renamed
