+++ tpm2_pytss/utils.py
